% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_adiabatic_lapse_rate_from_ct}
\alias{gsw_adiabatic_lapse_rate_from_ct}
\title{gsw_adiabatic_lapse_rate_from_ct}
\usage{
gsw_adiabatic_lapse_rate_from_ct(sa, ct, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the adiabatic lapse rate from Conservative Temperature
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_adiabatic_lapse_rate_from_ct.html}{Official TEOS-10 documentation for \code{gsw_adiabatic_lapse_rate_from_ct()}}
}
