% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_alpha}
\alias{gsw_alpha}
\title{gsw_alpha}
\usage{
gsw_alpha(sa, ct, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the thermal expansion coefficient of seawater with respect to
Conservative Temperature using the computationally-efficient 48-term
expression for density in terms of SA, CT and p (IOC et al., 2010)
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_alpha.html}{Official TEOS-10 documentation for \code{gsw_alpha()}}
}
