% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_alpha_on_beta}
\alias{gsw_alpha_on_beta}
\title{gsw_alpha_on_beta}
\usage{
gsw_alpha_on_beta(sa, ct, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates alpha divided by beta, where alpha is the thermal expansion
coefficient and beta is the saline contraction coefficient of seawater
from Absolute Salinity and Conservative Temperature.  This function uses
the computationally-efficient expression for specific volume in terms of
SA, CT and p (Roquet et al., 2014).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_alpha_on_beta.html}{Official TEOS-10 documentation for \code{gsw_alpha_on_beta()}}
}
