% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_cabbeling}
\alias{gsw_cabbeling}
\title{gsw_cabbeling}
\usage{
gsw_cabbeling(sa, ct, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the cabbeling coefficient of seawater with respect to
Conservative Temperature.  This function uses the computationally-
efficient expression for specific volume in terms of SA, CT and p
(Roquet et al., 2014).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_cabbeling.html}{Official TEOS-10 documentation for \code{gsw_cabbeling()}}
}
