% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_chem_potential_water_ice}
\alias{gsw_chem_potential_water_ice}
\title{gsw_chem_potential_water_ice}
\usage{
gsw_chem_potential_water_ice(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the chemical potential of water in ice from in-situ
temperature and pressure.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_chem_potential_water_ice.html}{Official TEOS-10 documentation for \code{gsw_chem_potential_water_ice()}}
}
