% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_ct_freezing}
\alias{gsw_ct_freezing}
\title{gsw_ct_freezing}
\usage{
gsw_ct_freezing(sa, p, saturation_fraction)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{p}{sea pressure \verb{[ dbar ]}}

\item{saturation_fraction}{the saturation fraction of dissolved air in seawater}
}
\description{
Calculates the Conservative Temperature at which seawater freezes.  The
Conservative Temperature freezing point is calculated from the exact
in-situ freezing temperature which is found by a modified Newton-Raphson
iteration (McDougall and Wotherspoon, 2013) of the equality of the
chemical potentials of water in seawater and in ice.
An alternative GSW function, gsw_CT_freezing_poly, it is based on a
computationally-efficient polynomial, and is accurate to within -5e-4 K
and 6e-4 K, when compared with this function.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_ct_freezing.html}{Official TEOS-10 documentation for \code{gsw_ct_freezing()}}
}
