% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_ct_freezing_poly}
\alias{gsw_ct_freezing_poly}
\title{gsw_ct_freezing_poly}
\usage{
gsw_ct_freezing_poly(sa, p, saturation_fraction)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{p}{sea pressure \verb{[ dbar ]}}

\item{saturation_fraction}{the saturation fraction of dissolved air in seawater}
}
\description{
Calculates the Conservative Temperature at which seawater freezes.
The error of this fit ranges between -5e-4 K and 6e-4 K when compared
with the Conservative Temperature calculated from the exact in-situ
freezing temperature which is found by a Newton-Raphson iteration of the
equality of the chemical potentials of water in seawater and in ice.
Note that the Conservative temperature freezing temperature can be found
by this exact method using the function gsw_CT_freezing.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_ct_freezing_poly.html}{Official TEOS-10 documentation for \code{gsw_ct_freezing_poly()}}
}
