% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_ct_from_enthalpy}
\alias{gsw_ct_from_enthalpy}
\title{gsw_ct_from_enthalpy}
\usage{
gsw_ct_from_enthalpy(sa, h, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{h}{specific enthalpy \verb{[ J/kg ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the Conservative Temperature of seawater, given the Absolute
Salinity, specific enthalpy, h, and pressure p.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_ct_from_enthalpy.html}{Official TEOS-10 documentation for \code{gsw_ct_from_enthalpy()}}
}
