% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_ct_from_pt}
\alias{gsw_ct_from_pt}
\title{gsw_ct_from_pt}
\usage{
gsw_ct_from_pt(sa, pt)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{pt}{potential temperature with reference pressure of 0 dbar \verb{[ deg C ]}}
}
\description{
Calculates Conservative Temperature from potential temperature of seawater
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_ct_from_pt.html}{Official TEOS-10 documentation for \code{gsw_ct_from_pt()}}
}
