% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_ct_maxdensity}
\alias{gsw_ct_maxdensity}
\title{gsw_ct_maxdensity}
\usage{
gsw_ct_maxdensity(sa, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the Conservative Temperature of maximum density of seawater.
This function returns the Conservative temperature at which the density
of seawater is a maximum, at given Absolute Salinity, SA, and sea
pressure, p (in dbar).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_ct_maxdensity.html}{Official TEOS-10 documentation for \code{gsw_ct_maxdensity()}}
}
