% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_deltasa_atlas}
\alias{gsw_deltasa_atlas}
\title{gsw_deltasa_atlas}
\usage{
gsw_deltasa_atlas(p, lon, lat)
}
\arguments{
\item{p}{sea pressure \verb{[ dbar ]}}

\item{lon}{longitude \verb{[ deg E ]}}

\item{lat}{latitude \verb{[ deg N ]}}
}
\description{
Calculates the Absolute Salinity Anomaly atlas value, delta_SA_atlas.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_deltasa_atlas.html}{Official TEOS-10 documentation for \code{gsw_deltasa_atlas()}}
}
