% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_dilution_coefficient_t_exact}
\alias{gsw_dilution_coefficient_t_exact}
\title{gsw_dilution_coefficient_t_exact}
\usage{
gsw_dilution_coefficient_t_exact(sa, t, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the dilution coefficient of seawater.  The dilution
coefficient of seawater is defined as the Absolute Salinity times the
second derivative of the Gibbs function with respect to Absolute
Salinity, that is, SA.*g_SA_SA.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_dilution_coefficient_t_exact.html}{Official TEOS-10 documentation for \code{gsw_dilution_coefficient_t_exact()}}
}
