% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_dynamic_enthalpy}
\alias{gsw_dynamic_enthalpy}
\title{gsw_dynamic_enthalpy}
\usage{
gsw_dynamic_enthalpy(sa, ct, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates dynamic enthalpy of seawater using the computationally-
efficient expression for specific volume in terms of SA, CT and p
(Roquet et al., 2014).  Dynamic enthalpy is defined as enthalpy minus
|  potential enthalpy (Young, 2010).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_dynamic_enthalpy.html}{Official TEOS-10 documentation for \code{gsw_dynamic_enthalpy()}}
}
