% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_enthalpy_ct_exact}
\alias{gsw_enthalpy_ct_exact}
\title{gsw_enthalpy_ct_exact}
\usage{
gsw_enthalpy_ct_exact(sa, ct, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates specific enthalpy of seawater from Absolute Salinity and
Conservative Temperature and pressure.
Note that this function uses the full Gibbs function.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_enthalpy_ct_exact.html}{Official TEOS-10 documentation for \code{gsw_enthalpy_ct_exact()}}
}
