% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_enthalpy_diff}
\alias{gsw_enthalpy_diff}
\title{gsw_enthalpy_diff}
\usage{
gsw_enthalpy_diff(sa, ct, p_shallow, p_deep)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p_shallow}{upper sea pressure (i.e. shallower absolute pressure - 10.1325 dbar) \verb{[ dbar ]}}

\item{p_deep}{lower sea pressure (i.e. deeper absolute pressure - 10.1325 dbar) \verb{[ dbar ]}}
}
\description{
Calculates the difference of the specific enthalpy of seawater between
two different pressures, p_deep (the deeper pressure) and p_shallow
(the shallower pressure), at the same values of SA and CT.  This
function uses the computationally-efficient expression for specific
volume in terms of SA, CT and p (Roquet et al., 2014).  The output
(enthalpy_diff_CT) is the specific enthalpy evaluated at (SA,CT,p_deep)
minus the specific enthalpy at (SA,CT,p_shallow).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_enthalpy_diff.html}{Official TEOS-10 documentation for \code{gsw_enthalpy_diff()}}
}
