% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_enthalpy_ice}
\alias{gsw_enthalpy_ice}
\title{gsw_enthalpy_ice}
\usage{
gsw_enthalpy_ice(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the specific enthalpy of ice (h_Ih).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_enthalpy_ice.html}{Official TEOS-10 documentation for \code{gsw_enthalpy_ice()}}
}
