% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_enthalpy_sso_0}
\alias{gsw_enthalpy_sso_0}
\title{gsw_enthalpy_sso_0}
\usage{
gsw_enthalpy_sso_0(p)
}
\arguments{
\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
This function calculates enthalpy at the Standard Ocean Salinity, SSO,
and at a Conservative Temperature of zero degrees C, as a function of
pressure, p, in dbar, using a streamlined version of the
computationally-efficient expression for specific volume, that is, a
streamlined version of the code "gsw_enthalpy(SA,CT,p)".
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_enthalpy_sso_0.html}{Official TEOS-10 documentation for \code{gsw_enthalpy_sso_0()}}
}
