% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_entropy_from_t}
\alias{gsw_entropy_from_t}
\title{gsw_entropy_from_t}
\usage{
gsw_entropy_from_t(sa, t, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the specific entropy of seawater
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_entropy_from_t.html}{Official TEOS-10 documentation for \code{gsw_entropy_from_t()}}
}
