% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_entropy_part}
\alias{gsw_entropy_part}
\title{gsw_entropy_part}
\usage{
gsw_entropy_part(sa, t, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
entropy minus the terms that are a function of only SA
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_entropy_part.html}{Official TEOS-10 documentation for \code{gsw_entropy_part()}}
}
