% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_entropy_part_zerop}
\alias{gsw_entropy_part_zerop}
\title{gsw_entropy_part_zerop}
\usage{
gsw_entropy_part_zerop(sa, pt0)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{pt0}{insitu temperature \verb{[ deg C ]}}
}
\description{
entropy part evaluated at the sea surface
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_entropy_part_zerop.html}{Official TEOS-10 documentation for \code{gsw_entropy_part_zerop()}}
}
