% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_fdelta}
\alias{gsw_fdelta}
\title{gsw_fdelta}
\usage{
gsw_fdelta(p, lon, lat)
}
\arguments{
\item{p}{sea pressure \verb{[ dbar ]}}

\item{lon}{longitude \verb{[ deg E ]}}

\item{lat}{latitude \verb{[ deg N ]}}
}
\description{
Calculates fdelta.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_fdelta.html}{Official TEOS-10 documentation for \code{gsw_fdelta()}}
}
