% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_gibbs_ice_part_t}
\alias{gsw_gibbs_ice_part_t}
\title{gsw_gibbs_ice_part_t}
\usage{
gsw_gibbs_ice_part_t(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
part of the the first temperature derivative of Gibbs energy of ice
that is the outout is gibbs_ice(1,0,t,p) + S0
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_gibbs_ice_part_t.html}{Official TEOS-10 documentation for \code{gsw_gibbs_ice_part_t()}}
}
