% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_gibbs_ice_pt0}
\alias{gsw_gibbs_ice_pt0}
\title{gsw_gibbs_ice_pt0}
\usage{
gsw_gibbs_ice_pt0(pt0)
}
\arguments{
\item{pt0}{insitu temperature \verb{[ deg C ]}}
}
\description{
Part of the the first temperature derivative of Gibbs energy of ice
that is the outout is "gibbs_ice(1,0,pt0,0) + s0"
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_gibbs_ice_pt0.html}{Official TEOS-10 documentation for \code{gsw_gibbs_ice_pt0()}}
}
