% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_gibbs_ice_pt0_pt0}
\alias{gsw_gibbs_ice_pt0_pt0}
\title{gsw_gibbs_ice_pt0_pt0}
\usage{
gsw_gibbs_ice_pt0_pt0(pt0)
}
\arguments{
\item{pt0}{insitu temperature \verb{[ deg C ]}}
}
\description{
The second temperature derivative of Gibbs energy of ice at the
potential temperature with reference sea pressure of zero dbar.  That is
the output is gibbs_ice(2,0,pt0,0).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_gibbs_ice_pt0_pt0.html}{Official TEOS-10 documentation for \code{gsw_gibbs_ice_pt0_pt0()}}
}
