% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_gibbs_pt0_pt0}
\alias{gsw_gibbs_pt0_pt0}
\title{gsw_gibbs_pt0_pt0}
\usage{
gsw_gibbs_pt0_pt0(sa, pt0)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{pt0}{insitu temperature \verb{[ deg C ]}}
}
\description{
gibbs_tt at (sa,pt,0)
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_gibbs_pt0_pt0.html}{Official TEOS-10 documentation for \code{gsw_gibbs_pt0_pt0()}}
}
