% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_grav}
\alias{gsw_grav}
\title{gsw_grav}
\usage{
gsw_grav(lat, p)
}
\arguments{
\item{lat}{latitude \verb{[ deg N ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates acceleration due to gravity as a function of latitude and as
a function of pressure in the ocean.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_grav.html}{Official TEOS-10 documentation for \code{gsw_grav()}}
}
