% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_helmholtz_energy_ice}
\alias{gsw_helmholtz_energy_ice}
\title{gsw_helmholtz_energy_ice}
\usage{
gsw_helmholtz_energy_ice(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the Helmholtz energy of ice.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_helmholtz_energy_ice.html}{Official TEOS-10 documentation for \code{gsw_helmholtz_energy_ice()}}
}
