% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_hill_ratio_at_sp2}
\alias{gsw_hill_ratio_at_sp2}
\title{gsw_hill_ratio_at_sp2}
\usage{
gsw_hill_ratio_at_sp2(t)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}
}
\description{
Calculates the Hill ratio, which is the adjustment needed to apply for
Practical Salinities smaller than 2.  This ratio is defined at a
Practical Salinity = 2 and in-situ temperature, t using PSS-78. The Hill
ratio is the ratio of 2 to the output of the Hill et al. (1986) formula
for Practical Salinity at the conductivity ratio, Rt, at which Practical
Salinity on the PSS-78 scale is exactly 2.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_hill_ratio_at_sp2.html}{Official TEOS-10 documentation for \code{gsw_hill_ratio_at_sp2()}}
}
