% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_internal_energy}
\alias{gsw_internal_energy}
\title{gsw_internal_energy}
\usage{
gsw_internal_energy(sa, ct, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates internal energy of seawater.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_internal_energy.html}{Official TEOS-10 documentation for \code{gsw_internal_energy()}}
}
