% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_kappa_const_t_ice}
\alias{gsw_kappa_const_t_ice}
\title{gsw_kappa_const_t_ice}
\usage{
gsw_kappa_const_t_ice(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates isothermal compressibility of ice.
Note. This is the compressibility of ice AT CONSTANT IN-SITU
TEMPERATURE
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_kappa_const_t_ice.html}{Official TEOS-10 documentation for \code{gsw_kappa_const_t_ice()}}
}
