% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_kappa_t_exact}
\alias{gsw_kappa_t_exact}
\title{gsw_kappa_t_exact}
\usage{
gsw_kappa_t_exact(sa, t, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
isentropic compressibility of seawater
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_kappa_t_exact.html}{Official TEOS-10 documentation for \code{gsw_kappa_t_exact()}}
}
