% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_latentheat_evap_ct}
\alias{gsw_latentheat_evap_ct}
\title{gsw_latentheat_evap_ct}
\usage{
gsw_latentheat_evap_ct(sa, ct)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}
}
\description{
Calculates latent heat, or enthalpy, of evaporation.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_latentheat_evap_ct.html}{Official TEOS-10 documentation for \code{gsw_latentheat_evap_ct()}}
}
