% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_latentheat_evap_t}
\alias{gsw_latentheat_evap_t}
\title{gsw_latentheat_evap_t}
\usage{
gsw_latentheat_evap_t(sa, t)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}
}
\description{
Calculates latent heat, or enthalpy, of evaporation.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_latentheat_evap_t.html}{Official TEOS-10 documentation for \code{gsw_latentheat_evap_t()}}
}
