% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_latentheat_melting}
\alias{gsw_latentheat_melting}
\title{gsw_latentheat_melting}
\usage{
gsw_latentheat_melting(sa, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates latent heat, or enthalpy, of melting.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_latentheat_melting.html}{Official TEOS-10 documentation for \code{gsw_latentheat_melting()}}
}
