% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_melting_ice_sa_ct_ratio_poly}
\alias{gsw_melting_ice_sa_ct_ratio_poly}
\title{gsw_melting_ice_sa_ct_ratio_poly}
\usage{
gsw_melting_ice_sa_ct_ratio_poly(sa, ct, p, t_ih)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}

\item{t_ih}{in-situ temperature of the ice at pressure p (ITS-90) \verb{[ deg C ]}}
}
\description{
Calculates the ratio of SA to CT changes when ice melts into seawater.
It is assumed that a small mass of ice melts into an infinite mass of
seawater.  Because of the infinite mass of seawater, the ice will always
melt.
The output, melting_seaice_SA_CT_ratio, is dSA/dCT rather than dCT/dSA.
This is done so that when SA = 0, the output, dSA/dCT is zero whereas
dCT/dSA would be infinite.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_melting_ice_sa_ct_ratio_poly.html}{Official TEOS-10 documentation for \code{gsw_melting_ice_sa_ct_ratio_poly()}}
}
