% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_melting_seaice_sa_ct_ratio_poly}
\alias{gsw_melting_seaice_sa_ct_ratio_poly}
\title{gsw_melting_seaice_sa_ct_ratio_poly}
\usage{
gsw_melting_seaice_sa_ct_ratio_poly(sa, ct, p, sa_seaice, t_seaice)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}

\item{sa_seaice}{Absolute Salinity of sea ice, that is, the mass fraction of salt in sea ice, expressed in g of salt per kg of sea ice. \verb{[ g/kg ]}}

\item{t_seaice}{the in-situ temperature of the sea ice (or ice) (ITS-90) \verb{[ deg C ]}}
}
\description{
Calculates the ratio of SA to CT changes when sea ice melts into seawater.
It is assumed that a small mass of sea ice melts into an infinite mass of
seawater.  Because of the infinite mass of seawater, the sea ice will
always melt.
Ice formed at the sea surface (sea ice) typically contains between 2 g/kg
and 12 g/kg of salt (defined as the mass of salt divided by the mass of
ice Ih plus brine) and this programme returns NaN's if the input
SA_seaice is greater than 15 g/kg.  If the SA_seaice input is not zero,
usually this would imply that the pressure p should be zero, as sea ice
only occurs near the sea surface.  The code does not impose that p = 0 if
SA_seaice is non-zero.  Rather, this is left to the user.
The Absolute Salinity, SA_brine, of the brine trapped in little pockets
in the sea ice, is in thermodynamic equilibrium with the ice Ih that
surrounds these pockets.  As the seaice temperature, t_seaice, may be
less than the freezing temperature, SA_brine is usually greater than the
Absolute Salinity of the seawater at the time and place when and where
the sea ice was formed.  So usually SA_brine will be larger than SA.
The output, melting_seaice_SA_CT_ratio, is dSA/dCT rather than dCT/dSA.
This is done so that when (SA - seaice_SA) = 0, the output, dSA/dCT is
zero whereas dCT/dSA would be infinite.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_melting_seaice_sa_ct_ratio_poly.html}{Official TEOS-10 documentation for \code{gsw_melting_seaice_sa_ct_ratio_poly()}}
}
