% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_Nsquared}
\alias{gsw_Nsquared}
\title{Calculate square of buoyancy frequency}
\usage{
gsw_Nsquared(SA, CT, p, latitude = 0)
}
\arguments{
\item{SA}{Absolute salinity.}

\item{CT}{Conservative temperature.}

\item{p}{Sea pressure in decibars.}

\item{latitude}{The latitude in deg North.}
}
\value{
Square of buoyancy frequency in 1/s^2, a vector of length 1 less than SA.
}
\description{
Calculate square of buoyancy frequency
}
\examples{
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.8099, 28.4392, 22.7862, 10.2262,  6.8272,  4.3236)
p <- c(      10,      50,     125,     250,     600,    1000)
latitude <- 4
gsw_Nsquared(SA, CT, p, latitude) # 1e-3*c(0.060846990523477, 0.235607737824943, 0.215533939997650, 0.012924024206854, 0.008425873682231)
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_Nsquared.html}
}

