% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_p_from_z}
\alias{gsw_p_from_z}
\title{gsw_p_from_z}
\usage{
gsw_p_from_z(z, lat)
}
\arguments{
\item{z}{height \verb{[ m ]}}

\item{lat}{latitude \verb{[ deg N ]}}
}
\description{
Calculates the pressure p from height z
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_p_from_z.html}{Official TEOS-10 documentation for \code{gsw_p_from_z()}}
}
