% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pot_enthalpy_from_pt_ice}
\alias{gsw_pot_enthalpy_from_pt_ice}
\title{gsw_pot_enthalpy_from_pt_ice}
\usage{
gsw_pot_enthalpy_from_pt_ice(pt0_ice)
}
\arguments{
\item{pt0_ice}{potential temperature of ice (ITS-90) \verb{[ deg C ]}}
}
\description{
Calculates the potential enthalpy of ice from potential temperature of
ice (whose reference sea pressure is zero dbar).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pot_enthalpy_from_pt_ice.html}{Official TEOS-10 documentation for \code{gsw_pot_enthalpy_from_pt_ice()}}
}
