% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pot_enthalpy_from_pt_ice_poly}
\alias{gsw_pot_enthalpy_from_pt_ice_poly}
\title{gsw_pot_enthalpy_from_pt_ice_poly}
\usage{
gsw_pot_enthalpy_from_pt_ice_poly(pt0_ice)
}
\arguments{
\item{pt0_ice}{potential temperature of ice (ITS-90) \verb{[ deg C ]}}
}
\description{
Calculates the potential enthalpy of ice from potential temperature of
ice (whose reference sea pressure is zero dbar).  This is a
compuationally efficient polynomial fit to the potential enthalpy of
ice.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pot_enthalpy_from_pt_ice_poly.html}{Official TEOS-10 documentation for \code{gsw_pot_enthalpy_from_pt_ice_poly()}}
}
