% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pot_enthalpy_ice_freezing}
\alias{gsw_pot_enthalpy_ice_freezing}
\title{gsw_pot_enthalpy_ice_freezing}
\usage{
gsw_pot_enthalpy_ice_freezing(sa, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the potential enthalpy of ice at which seawater freezes.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pot_enthalpy_ice_freezing.html}{Official TEOS-10 documentation for \code{gsw_pot_enthalpy_ice_freezing()}}
}
