% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pot_enthalpy_ice_freezing_poly}
\alias{gsw_pot_enthalpy_ice_freezing_poly}
\title{gsw_pot_enthalpy_ice_freezing_poly}
\usage{
gsw_pot_enthalpy_ice_freezing_poly(sa, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the potential enthalpy of ice at which seawater freezes.
The error of this fit ranges between -2.5 and 1 J/kg with an rms of
1.07, between SA of 0 and 120 g/kg and p between 0 and 10,000 dbar (the
error in the fit is between -0.7 and 0.7 with an rms of
0.3, between SA of 0 and 120 g/kg and p between 0 and 5,000 dbar) when
compared with the potential enthalpy calculated from the exact in-situ
freezing temperature which is found by a Newton-Raphson iteration of the
equality of the chemical potentials of water in seawater and in ice.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pot_enthalpy_ice_freezing_poly.html}{Official TEOS-10 documentation for \code{gsw_pot_enthalpy_ice_freezing_poly()}}
}
