% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pot_rho_t_exact}
\alias{gsw_pot_rho_t_exact}
\title{gsw_pot_rho_t_exact}
\usage{
gsw_pot_rho_t_exact(sa, t, p, p_ref)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}

\item{p_ref}{reference sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the potential density of seawater
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pot_rho_t_exact.html}{Official TEOS-10 documentation for \code{gsw_pot_rho_t_exact()}}
}
