% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pressure_freezing_ct}
\alias{gsw_pressure_freezing_ct}
\title{gsw_pressure_freezing_ct}
\usage{
gsw_pressure_freezing_ct(sa, ct, saturation_fraction)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{saturation_fraction}{the saturation fraction of dissolved air in seawater}
}
\description{
Calculates the pressure (in dbar) of seawater at the freezing
temperature.  That is, the output is the pressure at which seawater,
with Absolute Salinity SA, Conservative Temperature CT, and with
saturation_fraction of dissolved air, freezes.  If the input values are
such that there is no value of pressure in the range between 0 dbar and
10,000 dbar for which seawater is at the freezing temperature, the
output, pressure_freezing, is put equal to NaN.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pressure_freezing_ct.html}{Official TEOS-10 documentation for \code{gsw_pressure_freezing_ct()}}
}
