% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pt0_cold_ice_poly}
\alias{gsw_pt0_cold_ice_poly}
\title{gsw_pt0_cold_ice_poly}
\usage{
gsw_pt0_cold_ice_poly(pot_enthalpy_ice)
}
\arguments{
\item{pot_enthalpy_ice}{potential enthalpy of ice \verb{[ J/kg ]}}
}
\description{
Calculates an initial estimate of pt0_ice when it is less than about
-100 deg C.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pt0_cold_ice_poly.html}{Official TEOS-10 documentation for \code{gsw_pt0_cold_ice_poly()}}
}
