% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pt0_from_t_ice}
\alias{gsw_pt0_from_t_ice}
\title{gsw_pt0_from_t_ice}
\usage{
gsw_pt0_from_t_ice(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates potential temperature of ice Ih with a reference pressure of
0 dbar, from in-situ temperature, t.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pt0_from_t_ice.html}{Official TEOS-10 documentation for \code{gsw_pt0_from_t_ice()}}
}
