% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pt_from_ct}
\alias{gsw_pt_from_ct}
\title{gsw_pt_from_ct}
\usage{
gsw_pt_from_ct(sa, ct)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}
}
\description{
potential temperature of seawater from conservative temperature
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pt_from_ct.html}{Official TEOS-10 documentation for \code{gsw_pt_from_ct()}}
}
