% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pt_from_entropy}
\alias{gsw_pt_from_entropy}
\title{gsw_pt_from_entropy}
\usage{
gsw_pt_from_entropy(sa, entropy)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{entropy}{specific entropy \verb{[ deg C ]}}
}
\description{
Calculates potential temperature with reference pressure p_ref = 0 dbar
and with entropy as an input variable.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pt_from_entropy.html}{Official TEOS-10 documentation for \code{gsw_pt_from_entropy()}}
}
