% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pt_from_pot_enthalpy_ice}
\alias{gsw_pt_from_pot_enthalpy_ice}
\title{gsw_pt_from_pot_enthalpy_ice}
\usage{
gsw_pt_from_pot_enthalpy_ice(pot_enthalpy_ice)
}
\arguments{
\item{pot_enthalpy_ice}{potential enthalpy of ice \verb{[ J/kg ]}}
}
\description{
Calculates the potential temperature of ice from the potential enthalpy
of ice.  The reference sea pressure of both the potential temperature
and the potential enthalpy is zero dbar.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pt_from_pot_enthalpy_ice.html}{Official TEOS-10 documentation for \code{gsw_pt_from_pot_enthalpy_ice()}}
}
