% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pt_from_pot_enthalpy_ice_poly}
\alias{gsw_pt_from_pot_enthalpy_ice_poly}
\title{gsw_pt_from_pot_enthalpy_ice_poly}
\usage{
gsw_pt_from_pot_enthalpy_ice_poly(pot_enthalpy_ice)
}
\arguments{
\item{pot_enthalpy_ice}{potential enthalpy of ice \verb{[ J/kg ]}}
}
\description{
Calculates the potential temperature of ice (whose reference sea
pressure is zero dbar) from the potential enthalpy of ice.  This is a
compuationally efficient polynomial fit to the potential enthalpy of
ice.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pt_from_pot_enthalpy_ice_poly.html}{Official TEOS-10 documentation for \code{gsw_pt_from_pot_enthalpy_ice_poly()}}
}
