% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pt_from_pot_enthalpy_ice_poly_dh}
\alias{gsw_pt_from_pot_enthalpy_ice_poly_dh}
\title{gsw_pt_from_pot_enthalpy_ice_poly_dh}
\usage{
gsw_pt_from_pot_enthalpy_ice_poly_dh(pot_enthalpy_ice)
}
\arguments{
\item{pot_enthalpy_ice}{potential enthalpy of ice \verb{[ J/kg ]}}
}
\description{
Calculates the derivative of potential temperature of ice with respect
to potential enthalpy.  This is based on the compuationally-efficient
polynomial fit to the potential enthalpy of ice.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pt_from_pot_enthalpy_ice_poly_dh.html}{Official TEOS-10 documentation for \code{gsw_pt_from_pot_enthalpy_ice_poly_dh()}}
}
