% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pt_from_t}
\alias{gsw_pt_from_t}
\title{gsw_pt_from_t}
\usage{
gsw_pt_from_t(sa, t, p, p_ref)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}

\item{p_ref}{reference sea pressure \verb{[ dbar ]}}
}
\description{
Calculates potential temperature of seawater from in-situ temperature
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pt_from_t.html}{Official TEOS-10 documentation for \code{gsw_pt_from_t()}}
}
