% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_pt_from_t_ice}
\alias{gsw_pt_from_t_ice}
\title{gsw_pt_from_t_ice}
\usage{
gsw_pt_from_t_ice(t, p, p_ref)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}

\item{p_ref}{reference sea pressure \verb{[ dbar ]}}
}
\description{
Calculates potential temperature of ice Ih with the general reference
pressure, p_ref, from in-situ temperature, t.
A faster gsw routine exists if p_ref is indeed zero dbar.  This routine
is "gsw_pt0_from_t_ice(t,p)".
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_pt_from_t_ice.html}{Official TEOS-10 documentation for \code{gsw_pt_from_t_ice()}}
}
