% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_rho}
\alias{gsw_rho}
\title{gsw_rho}
\usage{
gsw_rho(sa, ct, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates in-situ density from Absolute Salinity and Conservative
Temperature, using the computationally-efficient expression for
specific volume in terms of SA, CT and p (Roquet et al., 2014).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_rho.html}{Official TEOS-10 documentation for \code{gsw_rho()}}
}
