% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_rho_ice}
\alias{gsw_rho_ice}
\title{gsw_rho_ice}
\usage{
gsw_rho_ice(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates in-situ density of ice from in-situ temperature and pressure.
Note that the output, rho_ice, is density, not density anomaly;  that
is, 1000 kg/m^3 is not subracted from it.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_rho_ice.html}{Official TEOS-10 documentation for \code{gsw_rho_ice()}}
}
