% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sa_freezing_from_ct}
\alias{gsw_sa_freezing_from_ct}
\title{gsw_sa_freezing_from_ct}
\usage{
gsw_sa_freezing_from_ct(ct, p, saturation_fraction)
}
\arguments{
\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}

\item{saturation_fraction}{the saturation fraction of dissolved air in seawater}
}
\description{
Calculates the Absolute Salinity of seawater at the freezing temperature.
That is, the output is the Absolute Salinity of seawater, with
Conservative Temperature CT, pressure p and the fraction
saturation_fraction of dissolved air, that is in equilibrium
with ice at the same in situ temperature and pressure.  If the input
values are such that there is no positive value of Absolute Salinity for
which seawater is frozen, the output is made a NaN.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sa_freezing_from_ct.html}{Official TEOS-10 documentation for \code{gsw_sa_freezing_from_ct()}}
}
