% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sa_freezing_from_t_poly}
\alias{gsw_sa_freezing_from_t_poly}
\title{gsw_sa_freezing_from_t_poly}
\usage{
gsw_sa_freezing_from_t_poly(t, p, saturation_fraction)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}

\item{saturation_fraction}{the saturation fraction of dissolved air in seawater}
}
\description{
Calculates the Absolute Salinity of seawater at the freezing temperature.
That is, the output is the Absolute Salinity of seawater, with the
fraction saturation_fraction of dissolved air, that is in equilibrium
with ice at in-situ temperature t and pressure p.  If the input values
are such that there is no positive value of Absolute Salinity for which
seawater is frozen, the output is put equal to Nan.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sa_freezing_from_t_poly.html}{Official TEOS-10 documentation for \code{gsw_sa_freezing_from_t_poly()}}
}
