% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sa_from_rho}
\alias{gsw_sa_from_rho}
\title{gsw_sa_from_rho}
\usage{
gsw_sa_from_rho(rho, ct, p)
}
\arguments{
\item{rho}{in-situ density    \verb{[ kg/m ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the Absolute Salinity of a seawater sample, for given values
of its density, Conservative Temperature and sea pressure (in dbar).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sa_from_rho.html}{Official TEOS-10 documentation for \code{gsw_sa_from_rho()}}
}
