% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sa_from_sp_baltic}
\alias{gsw_sa_from_sp_baltic}
\title{gsw_sa_from_sp_baltic}
\usage{
gsw_sa_from_sp_baltic(sp, lon, lat)
}
\arguments{
\item{sp}{Practical Salinity \verb{[ unitless ]}}

\item{lon}{longitude \verb{[ deg E ]}}

\item{lat}{latitude \verb{[ deg N ]}}
}
\description{
For the Baltic Sea, calculates Absolute Salinity with a value
computed analytically from Practical Salinity
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sa_from_sp_baltic.html}{Official TEOS-10 documentation for \code{gsw_sa_from_sp_baltic()}}
}
