% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sa_from_sstar}
\alias{gsw_sa_from_sstar}
\title{gsw_sa_from_sstar}
\usage{
gsw_sa_from_sstar(sstar, p, lon, lat)
}
\arguments{
\item{sstar}{Preformed Salinity \verb{[ g/kg ]}}

\item{p}{sea pressure \verb{[ dbar ]}}

\item{lon}{longitude \verb{[ deg E ]}}

\item{lat}{latitude \verb{[ deg N ]}}
}
\description{
Calculates Absolute Salinity, SA, from Preformed Salinity, Sstar.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sa_from_sstar.html}{Official TEOS-10 documentation for \code{gsw_sa_from_sstar()}}
}
