% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_saar}
\alias{gsw_saar}
\title{gsw_saar}
\usage{
gsw_saar(p, lon, lat)
}
\arguments{
\item{p}{sea pressure \verb{[ dbar ]}}

\item{lon}{longitude \verb{[ deg E ]}}

\item{lat}{latitude \verb{[ deg N ]}}
}
\description{
function gsw_saar(p,long,lat)
!==========================================================================
}
\details{
Calculates the Absolute Salinity Anomaly Ratio, SAAR at a geographic point.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_saar.html}{Official TEOS-10 documentation for \code{gsw_saar()}}
}
