% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sigma0}
\alias{gsw_sigma0}
\title{gsw_sigma0}
\usage{
gsw_sigma0(sa, ct)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}
}
\description{
Calculates potential density anomaly with reference pressure of 0 dbar,
this being this particular potential density minus 1000 kg/m^3.  This
function has inputs of Absolute Salinity and Conservative Temperature.
This function uses the computationally-efficient 48-term expression for
density in terms of SA, CT and p (IOC et al., 2010).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sigma0.html}{Official TEOS-10 documentation for \code{gsw_sigma0()}}
}
