% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sigma3}
\alias{gsw_sigma3}
\title{gsw_sigma3}
\usage{
gsw_sigma3(sa, ct)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}
}
\description{
Calculates potential density anomaly with reference pressure of 3000 dbar,
this being this particular potential density minus 1000 kg/m^3.  This
function has inputs of Absolute Salinity and Conservative Temperature.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sigma3.html}{Official TEOS-10 documentation for \code{gsw_sigma3()}}
}
