% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw_sa_from_sp.R
\name{gsw_SP_from_C}
\alias{gsw_SP_from_C}
\title{Convert from conductivity to practical salinity}
\usage{
gsw_SP_from_C(C, t, p)
}
\arguments{
\item{C}{Conductivity ratio. FIXME: check unit}

\item{t}{In-situ temperature in degrees C.}

\item{p}{Pressure in decibars. FIXME: check unit}
}
\value{
Practical salinity.
}
\description{
Convert from conductivity to practical salinity
}
\examples{
C <- c(34.5487, 34.7275, 34.8605, 34.6810, 34.5680, 34.5600)
t <- c(28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <- c(     10,      50,     125,     250,     600,    1000)
SP <- gsw_SP_from_C(C, t, p)
stopifnot(all.equal(SP,c(20.009869599086951,20.265511864874270,22.981513062527689,31.204503263727982,34.032315787432829,36.400308494388170)))
}

