% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sp_from_sa_baltic}
\alias{gsw_sp_from_sa_baltic}
\title{gsw_sp_from_sa_baltic}
\usage{
gsw_sp_from_sa_baltic(sa, lon, lat)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{lon}{longitude \verb{[ deg E ]}}

\item{lat}{latitude \verb{[ deg N ]}}
}
\description{
For the Baltic Sea, calculates Practical Salinity with a value
computed analytically from Absolute Salinity
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sp_from_sa_baltic.html}{Official TEOS-10 documentation for \code{gsw_sp_from_sa_baltic()}}
}
