% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sp_from_sk}
\alias{gsw_sp_from_sk}
\title{gsw_sp_from_sk}
\usage{
gsw_sp_from_sk(sk)
}
\arguments{
\item{sk}{Knudsen Salinity \verb{[ parts per thousand, ppt ]}}
}
\description{
Calculates Practical Salinity, SP, from SK
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sp_from_sk.html}{Official TEOS-10 documentation for \code{gsw_sp_from_sk()}}
}
