% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sp_from_sr}
\alias{gsw_sp_from_sr}
\title{gsw_sp_from_sr}
\usage{
gsw_sp_from_sr(sr)
}
\arguments{
\item{sr}{Reference Salinity \verb{[ g/kg ]}}
}
\description{
Calculates Practical Salinity, sp, from Reference Salinity, sr.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sp_from_sr.html}{Official TEOS-10 documentation for \code{gsw_sp_from_sr()}}
}
