% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_specvol}
\alias{gsw_specvol}
\title{gsw_specvol}
\usage{
gsw_specvol(sa, ct, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates specific volume from Absolute Salinity, Conservative
Temperature and pressure, using the computationally-efficient
polynomial expression for specific volume (Roquet et al., 2014).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_specvol.html}{Official TEOS-10 documentation for \code{gsw_specvol()}}
}
