% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_specvol_anom_standard}
\alias{gsw_specvol_anom_standard}
\title{gsw_specvol_anom_standard}
\usage{
gsw_specvol_anom_standard(sa, ct, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates specific volume anomaly of seawater.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_specvol_anom_standard.html}{Official TEOS-10 documentation for \code{gsw_specvol_anom_standard()}}
}
