% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_specvol_sso_0}
\alias{gsw_specvol_sso_0}
\title{gsw_specvol_sso_0}
\usage{
gsw_specvol_sso_0(p)
}
\arguments{
\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
This function calculates specific volume at the Standard Ocean Salinty,
SSO, and at a Conservative Temperature of zero degrees C, as a function
of pressure, p, in dbar, using a streamlined version of the CT version
of specific volume, that is, a streamlined version of the code
"gsw_specvol(SA,CT,p)".
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_specvol_sso_0.html}{Official TEOS-10 documentation for \code{gsw_specvol_sso_0()}}
}
