% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sr_from_sp}
\alias{gsw_sr_from_sp}
\title{gsw_sr_from_sp}
\usage{
gsw_sr_from_sp(sp)
}
\arguments{
\item{sp}{Practical Salinity \verb{[ unitless ]}}
}
\description{
Calculates Reference Salinity, SR, from Practical Salinity, SP.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sr_from_sp.html}{Official TEOS-10 documentation for \code{gsw_sr_from_sp()}}
}
