% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sstar_from_sa}
\alias{gsw_sstar_from_sa}
\title{gsw_sstar_from_sa}
\usage{
gsw_sstar_from_sa(sa, p, lon, lat)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{p}{sea pressure \verb{[ dbar ]}}

\item{lon}{longitude \verb{[ deg E ]}}

\item{lat}{latitude \verb{[ deg N ]}}
}
\description{
Calculates Preformed Salinity, Sstar, from Absolute Salinity, SA.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sstar_from_sa.html}{Official TEOS-10 documentation for \code{gsw_sstar_from_sa()}}
}
