% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sstar_from_sp}
\alias{gsw_sstar_from_sp}
\title{gsw_sstar_from_sp}
\usage{
gsw_sstar_from_sp(sp, p, lon, lat)
}
\arguments{
\item{sp}{Practical Salinity \verb{[ unitless ]}}

\item{p}{sea pressure \verb{[ dbar ]}}

\item{lon}{longitude \verb{[ deg E ]}}

\item{lat}{latitude \verb{[ deg N ]}}
}
\description{
Calculates Preformed Salinity, Sstar, from Practical Salinity, SP.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sstar_from_sp.html}{Official TEOS-10 documentation for \code{gsw_sstar_from_sp()}}
}
