% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_t_deriv_chem_potential_water_t_exact}
\alias{gsw_t_deriv_chem_potential_water_t_exact}
\title{gsw_t_deriv_chem_potential_water_t_exact}
\usage{
gsw_t_deriv_chem_potential_water_t_exact(sa, t, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the temperature derivative of the chemical potential of water
in seawater so that it is valid at exactly SA = 0.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_t_deriv_chem_potential_water_t_exact.html}{Official TEOS-10 documentation for \code{gsw_t_deriv_chem_potential_water_t_exact()}}
}
