% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_t_freezing_poly}
\alias{gsw_t_freezing_poly}
\title{gsw_t_freezing_poly}
\usage{
gsw_t_freezing_poly(sa, p, saturation_fraction)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{p}{sea pressure \verb{[ dbar ]}}

\item{saturation_fraction}{the saturation fraction of dissolved air in seawater}
}
\description{
Calculates the in-situ temperature at which seawater freezes from a
computationally efficient polynomial.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_t_freezing_poly.html}{Official TEOS-10 documentation for \code{gsw_t_freezing_poly()}}
}
