% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_t_from_pt0_ice}
\alias{gsw_t_from_pt0_ice}
\title{gsw_t_from_pt0_ice}
\usage{
gsw_t_from_pt0_ice(pt0_ice, p)
}
\arguments{
\item{pt0_ice}{potential temperature of ice (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates in-situ temperature from the potential temperature of ice Ih
with reference pressure, p_ref, of 0 dbar (the surface), and the
in-situ pressure.
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_t_from_pt0_ice.html}{Official TEOS-10 documentation for \code{gsw_t_from_pt0_ice()}}
}
