% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_thermobaric}
\alias{gsw_thermobaric}
\title{gsw_thermobaric}
\usage{
gsw_thermobaric(sa, ct, p)
}
\arguments{
\item{sa}{Absolute Salinity \verb{[ g/kg ]}}

\item{ct}{Conservative Temperature \verb{[deg C]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates the thermobaric coefficient of seawater with respect to
Conservative Temperature.  This routine is based on the
computationally-efficient expression for specific volume in terms of
SA, CT and p (Roquet et al., 2014).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_thermobaric.html}{Official TEOS-10 documentation for \code{gsw_thermobaric()}}
}
