% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_z_from_p}
\alias{gsw_z_from_p}
\title{gsw_z_from_p}
\usage{
gsw_z_from_p(p, lat)
}
\arguments{
\item{p}{sea pressure \verb{[ dbar ]}}

\item{lat}{latitude \verb{[ deg N ]}}
}
\description{
Calculates the height z from pressure p
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_z_from_p.html}{Official TEOS-10 documentation for \code{gsw_z_from_p()}}
}
