% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-package.R
\docType{package}
\name{teos10-package}
\alias{teos10}
\alias{teos10-package}
\title{teos10: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey.dunnington@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
