% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-wrappers.R
\name{gsw_sp_from_c}
\alias{gsw_sp_from_c}
\title{gsw_sp_from_c}
\usage{
gsw_sp_from_c(c, t, p)
}
\arguments{
\item{c}{conductivity \verb{[ mS/cm ]}}

\item{t}{in-situ temperature (ITS-90) \verb{[ deg C ]}}

\item{p}{sea pressure \verb{[ dbar ]}}
}
\description{
Calculates Practical Salinity, SP, from conductivity, C, primarily using
the PSS-78 algorithm.  Note that the PSS-78 algorithm for Practical
Salinity is only valid in the range 2 < SP < 42.  If the PSS-78
algorithm produces a Practical Salinity that is less than 2 then the
Practical Salinity is recalculated with a modified form of the Hill et
al. (1986) formula.  The modification of the Hill et al. (1986)
expression is to ensure that it is exactly consistent with PSS-78
at SP = 2.  Note that the input values of conductivity need to be in
units of mS/cm (not S/m).
}
\seealso{
\href{http://www.teos-10.org/pubs/gsw/html/gsw_sp_from_c.html}{Official TEOS-10 documentation for \code{gsw_sp_from_c()}}
}
