% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teos10-package.R
\docType{package}
\name{teos10-package}
\alias{teos10}
\alias{teos10-package}
\title{teos10: Zero Dependency Port of the TEOS-10 C API}
\description{
A complete, zero-dependenchy port of the TEOS-10
  seawater functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/teos10}
  \item Report bugs at \url{https://github.com/paleolimbot/teos10/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey.dunnington@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
